pinneaple-app/backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels>=4.0
daphne>=4.0
django-cors-headers>=4.3
python-decouple>=3.8
pillow>=10.0
numpy>=1.24
scipy>=1.10
